pyproject.toml
[project]
name = "radkit-mcp-server"
version = "1.0.0"
description = "FastMCP server for Cisco RADKit - Network device management via MCP protocol"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
keywords = ["cisco", "radkit", "mcp", "fastmcp", "network", "automation", "snmp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Networking",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "cisco-radkit-client==1.9.0",
    "cisco-radkit-common==1.9.0",
    "cisco-radkit-service==1.9.0",
    "fastmcp==2.13.1",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

onboarding = [
    "rich>=13.0.0",
    "questionary>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/CiscoDevNet/radkit-mcp-server-community"
Documentation = "https://github.com/CiscoDevNet/radkit-mcp-server-community#readme"
Repository = "https://github.com/CiscoDevNet/radkit-mcp-server-community"
Issues = "https://github.com/CiscoDevNet/radkit-mcp-server-community/issues"

[project.scripts]
radkit-mcp-server = "radkit_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/radkit_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/radkit_mcp"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
